Коммит с ДЗ № 8, (Stream) На проверку, с исправлением.
